mysite/requirements.txt
asgiref==3.4.1
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==35.0.0
Django==3.2.10
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-etc==1.3.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-recaptcha3==0.4.0
django-simple-captcha==0.5.14
django-taggit==2.0.0
idna==3.3
install==1.3.5
mysqlclient==2.1.0
Pillow==8.4.0
psycopg==3.0.7
psycopg-c==3.0.7
psycopg-pool==3.0.2
psycopg2-binary==2.9.2
pycparser==2.21
PyMySQL==1.0.2
pysha==0.0.2
pysha3==1.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
